Updated Android app with complete functionality and security fixes

+++ app/src/main/java/com/shoppino/android/data/repository/ProductRepository.kt
@@ -38,7 +38,7 @@ class ProductRepository(
         }
     }
     
-    suspend fun getProductById(id: Int): Product? {
+    suspend fun getProductById(id: Long): Product? {
         return productDao.getProductById(id)
     }
